+++ bridge/ChangeLog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+* Add an opt-in Kafka record envelope for JSON transformations
 
 ## Version 1.99.0
 * Include the Kafka partition in consumer idempotency keys